+++ .coveragerc
@@ -448,8 +448,11 @@ omit =
     homeassistant/components/zha/sensor.py
     homeassistant/components/zha/switch.py
     homeassistant/components/zha/api.py
-    homeassistant/components/zha/entities/*
-    homeassistant/components/zha/helpers.py
+    homeassistant/components/zha/entity.py
+    homeassistant/components/zha/device_entity.py
+    homeassistant/components/zha/core/helpers.py
+    homeassistant/components/zha/core/const.py
+    homeassistant/components/zha/core/listeners.py
     homeassistant/components/*/zha.py
 
     homeassistant/components/zigbee.py
