+++ homeassistant/components/zha/config_flow.py
@@ -6,9 +6,9 @@ import voluptuous as vol
 
 from homeassistant import config_entries
 
-from .const import (
+from .core.const import (
     CONF_RADIO_TYPE, CONF_USB_PATH, DEFAULT_DATABASE_NAME, DOMAIN, RadioType)
-from .helpers import check_zigpy_connection
+from .core.helpers import check_zigpy_connection
 
 
 @config_entries.HANDLERS.register(DOMAIN)
