+++ homeassistant/components/zha/core/__init__.py
@@ -1,10 +1,6 @@
 """
-Entities for Zigbee Home Automation.
+Core module for Zigbee Home Automation.
 
 For more details about this component, please refer to the documentation at
 https://home-assistant.io/components/zha/
 """
-
-# flake8: noqa
-from .device_entity import ZhaDeviceEntity
-from .entity import ZhaEntity
