+++ homeassistant/components/zha/core/helpers.py
