+++ homeassistant/components/zha/device_entity.py
