+++ homeassistant/components/zha/switch.py
@@ -8,10 +8,10 @@ import logging
 
 from homeassistant.components.switch import DOMAIN, SwitchDevice
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
-from . import helpers
-from .const import (
+from .core import helpers
+from .core.const import (
     DATA_ZHA, DATA_ZHA_DISPATCHERS, REPORT_CONFIG_IMMEDIATE, ZHA_DISCOVERY_NEW)
-from .entities import ZhaEntity
+from .entity import ZhaEntity
 
 _LOGGER = logging.getLogger(__name__)
